LayoutTests/imported/w3c/web-platform-tests/css/css-overflow/scroll-markers/scroll-marker-group-style-and-scroll-container-query-expected.txt


FAIL ::scroll-marker-group should not skip originating element for scroll-state queries assert_equals: expected "pass" but got ""
FAIL ::scroll-marker-group should not skip originating element for style queries assert_equals: expected "pass" but got ""
